- Made a fix to the decryptor that caused it to return the character that was one lower due to some rounding issue with double to char. - Implemented the decryption in the server/client and an encrypted login attempt can now be made.

+++ Common/Communication/Encryption/Encryptor.m
@@ -30,7 +30,7 @@ classdef Encryptor < handle
 					end
 				end
 			end
-			decrypted = char(decrypted);
+			decrypted = char(uint32(decrypted));
 		end
 	end
